update dark mode
