+++ src/components/Header/index.jsx
@@ -5,15 +5,18 @@
 
 import React from 'react';
 import { withTranslation } from 'react-i18next';
+import i18n from 'translations/i18n';
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { AUTHORIZATION_KEY, Storage } from 'aesirx-dma-lib';
-import { faQuestionCircle } from '@fortawesome/free-solid-svg-icons/faQuestionCircle';
+
 import { faChevronLeft } from '@fortawesome/free-solid-svg-icons/faChevronLeft';
+import { faGlobe } from '@fortawesome/free-solid-svg-icons/faGlobe';
 
 import './index.scss';
-
+import SwitchThemes from 'components/SwitchThemes';
 import DropdownAvatar from '../DropdownAvatar';
+import { Dropdown } from 'react-bootstrap';
 import Helper from '../../utils/helper';
 import ComponentHambuger from '../ComponentHambuger';
 import ComponentImage from '../ComponentImage';
@@ -39,8 +42,15 @@ class Header extends React.Component {
   };
 
   render() {
-    const { t } = this.props;
     let { isMini } = this.state;
+    const listLanguages = Object.keys(i18n.options.resources).map(function (key) {
+      return { language: key, title: i18n.options.resources[key].title };
+    });
+    let currentLanguage = listLanguages.filter((lang) => {
+      if (lang.language == i18n.language) {
+        return lang.title;
+      }
+    });
     return (
       <div
         id="all_header"
@@ -108,13 +118,38 @@ class Header extends React.Component {
                   <FontAwesomeIcon icon={faSearch} />
                 </button>
               </div> */}
-              <div className="wr_help_center ps-3 pe-3 d-none">
-                <span className="item_help d-flex align-items-center text-blue-0 cursor-pointer">
-                  <FontAwesomeIcon icon={faQuestionCircle} />
-                  <span className="text white-spacing-nowrap ps-2">{t('txt_help_center')}</span>
-                </span>
-              </div>
 
+              <div className="bottom-0 mb-2 py-1 button-language">
+                <Dropdown className="pt-2 ">
+                  <Dropdown.Toggle
+                    variant="dark"
+                    id="dropdown-basic"
+                    className="bg-transparent border-0"
+                  >
+                    <FontAwesomeIcon icon={faGlobe} /> {currentLanguage[0].title}
+                  </Dropdown.Toggle>
+
+                  <Dropdown.Menu>
+                    {listLanguages.map((item, index) => {
+                      return (
+                        <Dropdown.Item
+                          key={index}
+                          href="#"
+                          className=""
+                          onClick={() => {
+                            i18n.changeLanguage(item.language);
+                          }}
+                        >
+                          {item.title}
+                        </Dropdown.Item>
+                      );
+                    })}
+                  </Dropdown.Menu>
+                </Dropdown>
+              </div>
+              <div className="switch-theme-button ps-3 pe-3">
+                <SwitchThemes />
+              </div>
               <div className="ps-3 pe-3">
                 <DropdownAvatar />
               </div>
