+++ src/components/SbarLeft/index.jsx
@@ -5,15 +5,12 @@
 
 import React from 'react';
 import { withTranslation } from 'react-i18next';
-import i18n from 'translations/i18n';
 
 import './index.scss';
 import Menu from '../Menu';
 import Menu2 from '../Menu2';
-import SwitchThemes from 'components/SwitchThemes';
-import { Dropdown } from 'react-bootstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faGlobe } from '@fortawesome/free-solid-svg-icons/faGlobe';
+import { faQuestionCircle } from '@fortawesome/free-solid-svg-icons/faQuestionCircle';
 class SbarLeft extends React.Component {
   constructor(props) {
     super(props);
@@ -22,14 +19,7 @@ class SbarLeft extends React.Component {
 
   render() {
     let { settingPage } = this.props;
-    const listLanguages = Object.keys(i18n.options.resources).map(function (key) {
-      return { language: key, title: i18n.options.resources[key].title };
-    });
-    let currentLanguage = listLanguages.filter((lang) => {
-      if (lang.language == i18n.language) {
-        return lang.title;
-      }
-    });
+    const { t } = this.props;
     return (
       <aside
         className={`sidebar w-260  mt-0 position-relative bg-dark mh-100 overflow-hidden overflow-y-auto d-flex flex-column justify-content-between z-index-100`}
@@ -41,32 +31,13 @@ class SbarLeft extends React.Component {
         ) : (
           <Menu2 />
         )}
-        <div className="position-absolute bottom-0 mb-3 border-top w-100 py-1 button-language">
-          <Dropdown className="pt-2 ">
-            <Dropdown.Toggle variant="dark" id="dropdown-basic" className="bg-transparent border-0">
-              <FontAwesomeIcon icon={faGlobe} /> {currentLanguage[0].title}
-            </Dropdown.Toggle>
-
-            <Dropdown.Menu>
-              {listLanguages.map((item, index) => {
-                return (
-                  <Dropdown.Item
-                    key={index}
-                    href="#"
-                    className=""
-                    onClick={() => {
-                      i18n.changeLanguage(item.language);
-                    }}
-                  >
-                    {item.title}
-                  </Dropdown.Item>
-                );
-              })}
-            </Dropdown.Menu>
-          </Dropdown>
-        </div>
-        <div className="switch-theme-button position-absolute end-0 bottom-0 mb-4 me-3">
-          <SwitchThemes />
+        <div className="position-absolute bottom-0 me-3 border-top w-100 py-1">
+          <div className="wr_help_center ps-3 pe-3">
+            <span className="item_help d-flex align-items-center text-blue-0 cursor-pointer d-none">
+              <FontAwesomeIcon icon={faQuestionCircle} />
+              <span className="text white-spacing-nowrap ps-2">{t('txt_help_center')}</span>
+            </span>
+          </div>
         </div>
       </aside>
     );
