+++ src/themes/dark.scss
@@ -1,14 +1,14 @@
 [data-theme='dark'] {
-  --body-color: #bfc9f7;
-  --body-bg: #1a1a1a;
+  --body-color: #fff;
+  --body-bg: #0f1314;
 
   --box-shadow-color: rgb(26 26 26);
   --nav-bg: #1a2226;
 
-  --text-title-color: #c6c4f7;
-  --border-color: #373737;
+  --text-title-color: #fff;
+  --border-color: #3e545e;
 
-  --date-picker-color: #bfc9f7;
+  --date-picker-color: #fff;
   --date-picker-disabled-color: #282d34;
   --date-picker-header-color: #c6c4f7;
   --date-picker-selected: #373737;
@@ -21,35 +21,62 @@
   --input-placeholder-color: #282d34;
   --input-border-color: #373737;
 
-  --calendar-current-timeline: #3fff00;
-  --calendar-header-bg: #1d2c33;
-  --calendar-header-color: #bfc9f7;
-  --calendar-border-color: #1f2025;
-  --calendar-bg: #0d0d0d;
-  --calendar-color: #bfc9f7;
-  --calendar-outrange-bg: #0d0d0d;
-  --calendar-today-bg: #1f252c;
+  --calendar-current-timeline: #c60202;
+  --calendar-header-bg: #242d31;
+  --calendar-header-color: #fff;
+  --calendar-border-color: #3e545e;
+  --calendar-bg: #1a2226;
+  --calendar-color: #fff;
+  --calendar-outrange-bg: #1f252c;
+  --calendar-today-bg: #000c12;
+  --calendar-header-border: 1px solid #3e545e;
 
-  --categories-channels-color: #bfc9f7;
+  --categories-channels-color: #fff;
   --categories-channels-active-color: #0072a3;
 
   --dropdown-bg: #0d0d0d;
-  --dropdown-item-color: #bfc9f7;
-  --dropdown-item-hover-color: #bfc9f7;
+  --dropdown-item-color: #fff;
+  --dropdown-item-hover-color: #fff;
   --dropdown-item-hover-bg: #102129;
 
   --content-bg: #0d0d0d;
   --bg-blue-class: #1a1a1a;
   --view-active-bg: #3c3c3c;
   --text-black-class: #fff;
-  --text-blue-0-class: #bfc9f7;
+  --text-blue-0-class: #fff;
 
-  --event-bg: #001d49;
-  --event-color: #267cfd;
+  // --event-bg: #001d49;
+  // --event-color: #267cfd;
 
   --btn-green-color: #1a2226;
   --btn-outline-secondary-bg: #0d0d0d;
   --btn-outline-secondary-color: #ffffffcc;
   --btn-action-bg: #0d0d0d;
   --btn-action-color: #ffffffcc;
+
+  --joomla-calendar-bg: #c3e3de;
+  --youtube-calendar-bg: #eecacb;
+  --facebook-calendar-bg: #cfe2ff;
+  --twitter-calendar-bg: rgba(0, 172, 238, 0.2);
+  --linkedin-calendar-bg: rgba(0, 95, 137, 0.2);
+  --tumblr-calendar-bg: rgba(52, 58, 64, 0.2);
+  --medium-calendar-bg: rgba(14, 19, 24, 0.2);
+  --facebook-ads-calendar-bg: rgba(139, 186, 254, 0.2);
+  --google-adwords-calendar-bg: rgba(218, 165, 32, 0.2);
+  --wordpress-calendar-bg: rgba(214, 51, 132, 0.2);
+  --drupal-calendar-bg: rgba(70, 130, 180, 0.2);
+  --mailchimp-calendar-bg: rgba(139, 0, 139, 0.2);
+
+  --joomla-calendar-text: #16b979;
+  --youtube-calendar-text: #ef3737;
+  --facebook-calendar-text: #0d6efd;
+  --twitter-calendar-text: #00acee;
+  --linkedin-calendar-text: #005f89;
+  --tumblr-calendar-text: #343a40;
+  --medium-calendar-text: #0e1318;
+  --facebook-ads-calendar-text: #8bbafe;
+  --google-adwords-calendar-text: goldenrod;
+  --wordpress-calendar-text: #d63384;
+  --drupal-calendar-text: steelblue;
+  --mailchimp-calendar-text: darkmagenta;
 }
